src/requirements.txt
pyserial-asyncio
pynput==1.4.5